Phase3b完了: イベント処理分離とController層薄型化
- script.js: 382行 → 132行 (65%削減)
- eventHandlers.js: 364行の新規モジュール作成
- 全イベント処理をEventHandlersクラスに統合
- MVC設計の完全確立
- テスト168/168通過維持
